Collect phone numbers in feedback form. Closes #77
This change also changes the way we collect/update missing names.
Instead of using `current_member.update(user_attributes: { ... })` we
now use `current_member.user.update(...)` because, as it turns out, the
former has a tendency to _delete_ the current_member.user model, which
in turn deletes the current_member model (because of its destroy
dependency). This was _very_ bad and is now fixed.

This change also adds active_record_query_trace the development group in
the Gemfile. I used it to debug the deletion of current_member.user and
I think it'd be helpful to have the gem at our disposal in the future,
especially when we want to start thinking about performance
optimization.